Fix "-Wshorten-64-to-32" errors in Apple code
Also rename AndroidCallstack -> AppleCallstack in apple_callstack.cpp

+++ renderdoc/driver/metal/metal_common.cpp
@@ -703,15 +703,15 @@ static BlockShape GetBlockShape(MTL::PixelFormat mtlFormat, uint32_t plane)
 static uint32_t GetPlaneByteSize(uint32_t width, uint32_t height, uint32_t depth,
                                  MTL::PixelFormat mtlFormat, uint32_t mip, uint32_t plane)
 {
-  uint32_t mipWidth = RDCMAX(width >> mip, 1U);
-  uint32_t mipHeight = RDCMAX(height >> mip, 1U);
-  uint32_t mipDepth = RDCMAX(depth >> mip, 1U);
+  uint32_t mipWidth = RDCMAX(width >> mip, 1u);
+  uint32_t mipHeight = RDCMAX(height >> mip, 1u);
+  uint32_t mipDepth = RDCMAX(depth >> mip, 1u);
 
   MTL::Size planeShape(mipWidth, mipHeight, 0);
   BlockShape blockShape(GetBlockShape(mtlFormat, plane));
 
-  uint32_t widthInBlocks = (planeShape.width + blockShape.width - 1) / blockShape.width;
-  uint32_t heightInBlocks = (planeShape.height + blockShape.height - 1) / blockShape.height;
+  uint32_t widthInBlocks = (uint32_t)(planeShape.width + blockShape.width - 1) / blockShape.width;
+  uint32_t heightInBlocks = (uint32_t)(planeShape.height + blockShape.height - 1) / blockShape.height;
 
   return blockShape.bytes * widthInBlocks * heightInBlocks * mipDepth;
 }
